Initial version, experiment.
there are 3 possible ways to decorate service
1) @injector.inject each method explicitly. Will work but doesn't support
request-scope without requiring clients of the lib to add dependency provider
manually.
2) @injector.inject service class. This way the class will get dependency
provider authomatically. The provider will be mocked during the testing so
the we need to explore it better.
3) Extend custom class that decorates the methods and adds the provider.

Another thing to consider is a caching of dependencies in the request scope (if
needed). It is required in the case when resource creation is expensive and
there are several injection points.