Fix request scope
Problem
-------
Provider result in scope of the request is not cached therefored each time we try to
access an object we get a new instance of it when callable provider is used.
It's not a desired behaviour. Instead we should get the same object regardless
how many times it's injected.

Decision
--------
Use eventlet.corolocal storage. It's like threadlocal but for coroutines.

Testing
-------
Added test cases to ensure that we inject the same objet that is bound in
request-scope.