Add new type of scope to allow closing resources
Problem
-------
Some resources in request scope need to be closed (freed). It's not currently
supported.

Solution
--------
Create a new type of scope that's aware about resources with 'close' method.
Name of this special method is hardcoded. Probably we could use protocol in this
case - `Closable`.

It's a separate type of scope to make the fact that resource will be closed more
explicit. When one of the resources cannot be closed, the failure will be logged
but teardown call is not failed so the rest can be closed.

**Second change** is that ServiceContainer is now in singleton scope as it
should be.

Testing
-------
Added test cases.