Provide werkzeug.wrappers.Request as dependency by default
Context
-------
It's so common that other dependencies depend on HTTP request object that we
need to provide one by default. For instance other dependencies may take some
data from the headers of the request or arguments.

Solution
--------
Bind the request instance to werkzeug's class that is used by nameko.
Do it right in the nameko provider - no need to build a provider that depends on
nameko worker context etc.

The only problem is parent cannot access child dependencies so I need to rebind
them all in child injector.

Testing
-------
Test case provided.