Provide werkzeug.wrappers.Request as dependency by default (#5)
Context
-------
It's so common that other dependencies depend on HTTP request object that we
need to provide one by default. For instance other dependencies may take some
data from the headers of the request or arguments.

Solution
--------
Bind the request instance to werkzeug's class that is used by nameko.
Do it right in the nameko provider - no need to build a provider that depends on
nameko worker context etc.

The only problem was that with my current approach using parent-child injector - it doesn't work. All the providers defined on parent that depend on request cannot access that request, because parent injector doesn't know about children.

Now we don't create child injector but instead put the current request directly in the scope, same happens with worker_context. Only when testing test_injector we use child injector.

Testing
-------
Test case provided.